+++ Reach/src/utopia/reach/container/wrapper/AlignFrame.scala
@@ -1,37 +1,53 @@
 package utopia.reach.container.wrapper
 
 import utopia.firmament.component.container.single.AlignFrameLike
-import utopia.firmament.drawing.immutable.BackgroundDrawer
+import utopia.firmament.drawing.immutable.CustomDrawableFactory
 import utopia.firmament.drawing.template.CustomDrawer
-import utopia.paradigm.color.Color
-import utopia.paradigm.enumeration.Alignment
+import utopia.paradigm.enumeration.{Alignment, FromAlignmentFactory}
 import utopia.reach.component.factory.ComponentFactoryFactory.Cff
-import utopia.reach.component.factory.FromContextComponentFactoryFactory.Ccff
-import utopia.reach.component.factory.{BuilderFactory, ComponentFactoryFactory, FromGenericContextFactory, GenericContextualFactory, SimpleFilledBuilderFactory}
+import utopia.reach.component.factory.{FromGenericContextFactory, GenericContextualFactory}
 import utopia.reach.component.hierarchy.ComponentHierarchy
 import utopia.reach.component.template.{CustomDrawReachComponent, ReachComponentLike}
-import utopia.reach.component.wrapper.{ComponentCreationResult, Open, OpenComponent}
-import utopia.reach.container.ReachCanvas
+import utopia.reach.component.wrapper.{ComponentWrapResult, OpenComponent}
 
 object AlignFrame extends Cff[AlignFrameFactory]
 {
 	override def apply(hierarchy: ComponentHierarchy) = new AlignFrameFactory(hierarchy)
 }
 
+trait AlignFrameFactoryLike[+Repr]
+	extends WrapperContainerFactory[AlignFrame, ReachComponentLike] with CustomDrawableFactory[Repr]
+{
+	// ABSTRACT --------------------------
+	
+	/**
+	  * @return The alignment used in this factory
+	  */
+	def alignment: Alignment
+	
+	
+	// IMPLEMENTED  ---------------------
+	
+	override def apply[C <: ReachComponentLike, R](content: OpenComponent[C, R]): ComponentWrapResult[AlignFrame, C, R] = {
+		val frame = new AlignFrame(parentHierarchy, content.component, alignment, customDrawers)
+		content attachTo frame
+	}
+}
+
 class AlignFrameFactory(val parentHierarchy: ComponentHierarchy)
-	extends FromGenericContextFactory[Any, ContextualAlignFrameFactory] with BuilderFactory[AlignFrameBuilder]
-		with SimpleFilledBuilderFactory[ContextualFilledAlignFrameBuilder]
+	extends FromGenericContextFactory[Any, ContextualAlignFrameFactory]
+		with FromAlignmentFactory[InitializedAlignFrameFactory]
 {
 	// IMPLEMENTED  ------------------------------
 	
+	/**
+	  * @param alignment The alignment to use
+	  * @return A factory that uses the specified alignment
+	  */
+	override def apply(alignment: Alignment) = InitializedAlignFrameFactory(parentHierarchy, alignment)
+	
 	override def withContext[N <: Any](context: N) =
-		ContextualAlignFrameFactory(this, context)
-	
-	override def build[FF](contentFactory: ComponentFactoryFactory[FF]) =
-		new AlignFrameBuilder[FF](this, contentFactory)
-	
-	protected def makeBuilder[NC, F](background: Color, contentContext: NC, contentFactory: Ccff[NC, F]) =
-		new ContextualFilledAlignFrameBuilder[NC, F](this, background, contentContext, contentFactory)
+		ContextualAlignFrameFactory(parentHierarchy, context)
 	
 	
 	// OTHER    ----------------------------------
@@ -45,6 +61,7 @@ class AlignFrameFactory(val parentHierarchy: ComponentHierarchy)
 	 * @tparam R Type of additional creation result
 	 * @return A new align frame
 	 */
+	@deprecated("Please use .apply(Alignment).apply(OpenComponent) instead", "v1.1")
 	def apply[C <: ReachComponentLike, R](content: OpenComponent[C, R], alignment: Alignment,
 	                                      customDrawers: Vector[CustomDrawer] = Vector()) =
 	{
@@ -53,95 +70,51 @@ class AlignFrameFactory(val parentHierarchy: ComponentHierarchy)
 	}
 }
 
-case class ContextualAlignFrameFactory[N](factory: AlignFrameFactory, context: N)
+case class InitializedAlignFrameFactory(parentHierarchy: ComponentHierarchy, alignment: Alignment,
+                                        customDrawers: Vector[CustomDrawer] = Vector())
+	extends AlignFrameFactoryLike[InitializedAlignFrameFactory]
+		with NonContextualWrapperContainerFactory[AlignFrame, ReachComponentLike]
+		with FromGenericContextFactory[Any, InitializedContextualAlignFrameFactory]
+{
+	override def withCustomDrawers(drawers: Vector[CustomDrawer]): InitializedAlignFrameFactory =
+		copy(customDrawers = drawers)
+	
+	override def withContext[N <: Any](context: N): InitializedContextualAlignFrameFactory[N] =
+		InitializedContextualAlignFrameFactory(parentHierarchy, context, alignment, customDrawers)
+}
+
+case class ContextualAlignFrameFactory[N](parentHierarchy: ComponentHierarchy, context: N)
 	extends GenericContextualFactory[N, Any, ContextualAlignFrameFactory]
+		with FromAlignmentFactory[InitializedContextualAlignFrameFactory[N]]
 {
 	// COMPUTED ------------------------------
 	
 	/**
 	 * @return A copy of this factory with no contextual information
 	 */
-	def withoutContext = factory
+	@deprecated("Deprecated for removal", "v1.1")
+	def withoutContext = new AlignFrameFactory(parentHierarchy)
 	
 	
 	// IMPLEMENTED  --------------------------
 	
+	override def apply(alignment: Alignment) =
+		InitializedContextualAlignFrameFactory(parentHierarchy, context, alignment)
+	
 	override def withContext[N2 <: Any](newContext: N2) = copy(context = newContext)
-	
-	
-	// OTHER    ------------------------------
-	
-	/**
-	 * Creates a new builder based on this factory
-	 * @param contentFactory A frame content creation factory factory
-	 * @tparam F Type of component creation factory used
-	 * @return A new builder
-	 */
-	def build[F](contentFactory: Ccff[N, F]) = new ContextualAlignFrameBuilder(factory, context, contentFactory)
 }
 
-class AlignFrameBuilder[+F](factory: AlignFrameFactory, contentFactory: ComponentFactoryFactory[F])
+case class InitializedContextualAlignFrameFactory[N](parentHierarchy: ComponentHierarchy, context: N,
+                                                     alignment: Alignment,
+                                                     customDrawers: Vector[CustomDrawer] = Vector())
+	extends AlignFrameFactoryLike[InitializedContextualAlignFrameFactory[N]]
+		with ContextualWrapperContainerFactory[N, Any, AlignFrame, ReachComponentLike, InitializedContextualAlignFrameFactory]
 {
-	private implicit def canvas: ReachCanvas = factory.parentHierarchy.top
+	override def withContext[N2 <: Any](newContext: N2): InitializedContextualAlignFrameFactory[N2] =
+		copy(context = newContext)
 	
-	/**
-	 * Creates a new filled align frame
-	 * @param alignment Alignment to use
-	 * @param customDrawers Custom drawers to assign (default = empty)
-	 * @param fill A function for creating container contents. Accepts component creation factory.
-	 * @tparam C Type of wrapped component
-	 * @tparam R Type of additional creation result
-	 * @return A new align frame
-	 */
-	def apply[C <: ReachComponentLike, R](alignment: Alignment, customDrawers: Vector[CustomDrawer] = Vector())
-	                                     (fill: F => ComponentCreationResult[C, R]) =
-	{
-		val content = Open.using(contentFactory)(fill)
-		factory(content, alignment, customDrawers)
-	}
-}
-
-class ContextualAlignFrameBuilder[N, +F](factory: AlignFrameFactory, context: N, contentFactory: Ccff[N, F])
-{
-	private implicit def canvas: ReachCanvas = factory.parentHierarchy.top
-	
-	/**
-	 * Creates a new filled align frame
-	 * @param alignment Alignment to use
-	 * @param customDrawers Custom drawers to assign (default = empty)
-	 * @param fill A function for creating container contents. Accepts component creation factory.
-	 * @tparam C Type of wrapped component
-	 * @tparam R Type of additional creation result
-	 * @return A new align frame
-	 */
-	def apply[C <: ReachComponentLike, R](alignment: Alignment, customDrawers: Vector[CustomDrawer] = Vector())
-	                                     (fill: F => ComponentCreationResult[C, R]) =
-	{
-		val content = Open.withContext(context)(contentFactory)(fill)
-		factory(content, alignment, customDrawers)
-	}
-}
-
-class ContextualFilledAlignFrameBuilder[NC, +F](factory: AlignFrameFactory, background: Color, contentContext: NC,
-                                                contentFactory: Ccff[NC, F])
-{
-	private implicit def canvas: ReachCanvas = factory.parentHierarchy.top
-	
-	/**
-	 * Creates a new filled align frame
-	 * @param alignment Alignment to use
-	 * @param customDrawers Custom drawers to assign (default = empty)
-	 * @param fill A function for creating container contents. Accepts component creation factory.
-	 * @tparam C Type of wrapped component
-	 * @tparam R Type of additional creation result
-	 * @return A new align frame
-	 */
-	def apply[C <: ReachComponentLike, R](alignment: Alignment, customDrawers: Vector[CustomDrawer] = Vector())
-	                                     (fill: F => ComponentCreationResult[C, R]) =
-	{
-		val content = Open.withContext(contentContext)(contentFactory)(fill)
-		factory(content, alignment, BackgroundDrawer(background) +: customDrawers)
-	}
+	override def withCustomDrawers(drawers: Vector[CustomDrawer]): InitializedContextualAlignFrameFactory[N] =
+		copy(customDrawers = drawers)
 }
 
 /**